requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.128.0
fonttools==4.61.1
fpdf2==2.8.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.9
mangum==0.21.0
matplotlib==3.10.8
numpy==2.4.2
openai==2.16.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pandas-flavor==0.8.1
patsy==1.0.2
pillow==12.1.0
pingouin==0.5.5
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
statsmodels==0.14.6
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
xarray==2026.1.0
